Improved command line parameters.

+++ scripts/release.sh
@@ -45,6 +45,13 @@ then
     exit 1
 fi
 
+Release/storm/StormMain.exe -c bs("Hello",2){,[.,]} > /dev/null
+if [[ $? -ne 0 ]]
+then
+    echo "Failed to execute BS code. Test in Release/storm"
+    exit 1
+fi
+
 rm -r Release/storm
 
 echo "All seems well, copying the zip to fprg.se..."
